+++ gendc_python/__init__.py
@@ -1,1 +1,2 @@
 from . import gendc_separator
+from . import genicam
