+++ gendc_python/gendc_separator/__init__.py
@@ -1,2 +1,1 @@
 from .descriptor import Container
-from .tools import convert_pixelformat
