+++ gendc_python/genicam/__init__.py
@@ -1,0 +1,1 @@
+from .pfnc_converter import convert_pixelformat
