+++ gendc_python/genicam/data_str_key.py
